backend/requirements.txt
google-genai
fastapi[standard]
pydantic
python-dotenv
uvicorn
python-multipart
pillow